Move portfolio section logic to new file
